+++ demisto_sdk/commands/common/tools.py
@@ -14,7 +14,6 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Type, Union
 
 import click
 import colorama
-import demisto_sdk.commands.common.tools as tools
 import git
 import requests
 import urllib3
@@ -39,7 +38,7 @@ urllib3.disable_warnings()
 colorama.init()
 
 ryaml = YAML()
-ryaml.preserve_quotes = True  # type: ignore
+ryaml.preserve_quotes = True
 ryaml.allow_duplicate_keys = True
 
 
@@ -296,8 +295,10 @@ def get_last_remote_release_version():
         except Exception as exc:
             exc_msg = str(exc)
             if isinstance(exc, requests.exceptions.ConnectionError):
-                exc_msg = f'{exc_msg[exc_msg.find(">") + 3:-3]}.\nThis may happen if you are not connected to the internet.'
+                exc_msg = f'{exc_msg[exc_msg.find(">") + 3:-3]}.\n' \
+                          f'This may happen if you are not connected to the internet.'
             print_warning(f'Could not get latest demisto-sdk version.\nEncountered error: {exc_msg}')
+
     return ''
 
 
@@ -580,7 +581,17 @@ def is_file_path_in_pack(file_path):
 
 
 def get_pack_name(file_path):
-    match = re.search(r'^(?:./)?{}/([^/]+)/'.format(PACKS_DIR), file_path)
+    """
+    extract pack name (folder name) from file path
+
+    Arguments:
+        file_path (str): path of a file inside the pack
+
+    Returns:
+        pack name (str)
+    """
+    # the regex extracts pack name from relative paths, for example: Packs/EWSv2 -> EWSv2
+    match = re.search(rf'^{PACKS_DIR_REGEX}[/\\]([^/\\]+)[/\\]', file_path)
     return match.group(1) if match else None
 
 
@@ -737,7 +748,7 @@ def get_dict_from_file(path: str, use_ryaml: bool = False) -> Tuple[Dict, Union[
     return {}, None
 
 
-def find_type(path: str = '', _dict=None, file_type: Optional[str] = None):  # noqa: C901
+def find_type(path: str = '', _dict=None, file_type: Optional[str] = None, ignore_sub_categories: bool = False):  # noqa: C901
     """
     returns the content file type
 
@@ -770,13 +781,13 @@ def find_type(path: str = '', _dict=None, file_type: Optional[str] = None):  # n
 
     if file_type == 'yml':
         if 'category' in _dict:
-            if 'beta' in _dict:
+            if 'beta' in _dict and not ignore_sub_categories:
                 return FileType.BETA_INTEGRATION
 
             return FileType.INTEGRATION
 
         if 'script' in _dict:
-            if TEST_PLAYBOOKS_DIR in path:
+            if TEST_PLAYBOOKS_DIR in path and not ignore_sub_categories:
                 return FileType.TEST_SCRIPT
 
             return FileType.SCRIPT
@@ -878,7 +889,7 @@ def get_content_path() -> str:
         git_repo = git.Repo(os.getcwd(), search_parent_directories=True)
         remote_url = git_repo.remote().urls.__next__()
         is_fork_repo = 'content' in remote_url
-        is_external_repo = tools.is_external_repository()
+        is_external_repo = is_external_repository()
 
         if not is_fork_repo and not is_external_repo:
             raise git.InvalidGitRepositoryError
@@ -973,7 +984,7 @@ def is_file_from_content_repo(file_path: str) -> Tuple[bool, str]:
                         search_parent_directories=True)
     remote_url = git_repo.remote().urls.__next__()
     is_fork_repo = 'content' in remote_url
-    is_external_repo = tools.is_external_repository()
+    is_external_repo = is_external_repository()
 
     if not is_fork_repo and not is_external_repo:
         return False, ''
@@ -1045,7 +1056,6 @@ def is_test_config_match(test_config: dict, test_playbook_id: str = '', integrat
     If file type is not an integration- will return True if the test_playbook id matches playbookID.
     Args:
         test_config: A test configuration from conf.json file under 'tests' key.
-        file_type: The file type. can be 'integration', 'playbook'.
         test_playbook_id: A test playbook ID.
         integration_id: An integration ID.
     If both test_playbook_id and integration_id are given will look for a match of both, else will look for match